backend/zoverflow-image/seo_image/requirements.txt
Pillow~=9.4.0
boto3~=1.26.59